added get customer request based on auth header
