+++ src/main/java/com/knitted/marketplace/services/CustomerService.java
@@ -1,17 +1,23 @@
 package com.knitted.marketplace.services;
 
 import com.knitted.marketplace.dtos.auth.RegistrationRequestDto;
+import com.knitted.marketplace.exception.exceptions.RecordNotFoundException;
 import com.knitted.marketplace.mappers.AuthMapper;
 import com.knitted.marketplace.models.Customer;
 import com.knitted.marketplace.repositories.CustomerRepository;
+import com.knitted.marketplace.security.JwtService;
+import com.knitted.marketplace.utils.Parser;
 import org.springframework.stereotype.Service;
 
 @Service
 public class CustomerService {
-    private final CustomerRepository customerRepository;
 
-    public CustomerService(CustomerRepository customerRepository) {
+    private final CustomerRepository customerRepository;
+    private final JwtService jwtService;
+
+    public CustomerService(CustomerRepository customerRepository, JwtService jwtService) {
         this.customerRepository = customerRepository;
+        this.jwtService = jwtService;
     }
 
     public Customer createCustomer(RegistrationRequestDto request) {
@@ -20,5 +26,11 @@ public class CustomerService {
 
     public Customer saveCustomer(Customer customer) {
         return customerRepository.save(customer);
+    }
+
+    public Customer getCustomerByAuthHeader(String authHeader) {
+        String token = Parser.toToken(authHeader);
+        Long id = jwtService.extractId(token);
+        return customerRepository.findById(id).orElseThrow(() -> new RecordNotFoundException(id));
     }
 }
